refactor(checkbox): move component and story file from tailwind folder
