+++ tailwind/ui/Switch.tsx
@@ -3,7 +3,7 @@ import * as SwitchPrimitives from "@radix-ui/react-switch"
 
 import { cn } from "@/lib/utils/cn"
 
-import { commonControlClasses } from "./Checkbox"
+import { commonControlClasses } from "../../src/components/ui/checkbox"
 
 const Switch = React.forwardRef<
   React.ElementRef<typeof SwitchPrimitives.Root>,
